Extract state and doctor command services
